requirements.txt
# Core dependencies
garminconnect==0.2.8
duckdb==1.1.3
pandas==2.2.3
pyarrow==18.1.0
tabulate==0.9.0

# Web app
streamlit==1.40.1
plotly==5.24.1

# LLM
google-generativeai==0.8.3

# Environment & utilities
python-dotenv==1.0.1